backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
SQLAlchemy>=1.4
pydantic>=1.10,<2
python-multipart>=0.0.9
pymysql>=1.1
cryptography>=41.0.0
paho-mqtt>=1.6.1
openai>=1.0.0
python-dotenv>=1.0.0
